backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
pillow==10.0.0
spacy==3.6.1
requests==2.31.0
